refactor(audit): split review/render and engine/invocation (#5645)
