+++ src/core/engine/invocation.rs
@@ -5,9 +5,6 @@ use crate::core::error::{Error, Result};
 use crate::core::paths;
 use serde::{Deserialize, Serialize};
 use std::fs;
-use std::io;
-#[cfg(unix)]
-use std::os::unix::fs::MetadataExt;
 use std::path::{Path, PathBuf};
 use std::thread;
 use std::time::{Duration, SystemTime};
@@ -20,6 +17,7 @@ const PORT_POOL_START: u16 = 20_000;
 const PORT_POOL_END: u16 = 60_999;
 
 mod child;
+mod errors;
 mod runtime;
 pub use child::{
     cleanup_invocation_children, cleanup_stale_child_records, register_child_process,
@@ -135,7 +133,7 @@ impl InvocationGuard {
 
         for dir in [&state_dir, &artifact_dir, &tmp_dir] {
             fs::create_dir_all(dir)
-                .map_err(|e| invocation_dir_create_error(dir, &runtime_root, e))?;
+                .map_err(|e| errors::invocation_dir_create_error(dir, &runtime_root, e))?;
         }
 
         let needs_lease =
@@ -279,60 +277,6 @@ impl InvocationGuard {
             crate::core::artifact_manifest::manifest_for_existing_files(target)?
         };
         crate::core::artifact_manifest::write_manifest_to_root(target, &manifest)
-    }
-}
-
-fn invocation_dir_create_error(dir: &Path, runtime_root: &Path, error: io::Error) -> Error {
-    if error.kind() == io::ErrorKind::PermissionDenied {
-        return Error::internal_unexpected(format!(
-            "Homeboy cannot create invocation dir {} because runtime temp root {} is not writable by the current user ({}). {} Remediation: remove or chown the runtime temp root, or set {} to a writable short path.",
-            dir.display(),
-            runtime_root.display(),
-            current_user_label(),
-            runtime_root_owner_label(runtime_root),
-            HOMEBOY_INVOCATION_RUNTIME_DIR_ENV,
-        ));
-    }
-
-    Error::internal_io(
-        format!(
-            "Failed to create invocation dir {}: {}",
-            dir.display(),
-            error
-        ),
-        Some("invocation.dir".to_string()),
-    )
-}
-
-fn current_user_label() -> String {
-    let name = std::env::var("USER")
-        .or_else(|_| std::env::var("USERNAME"))
-        .ok()
-        .filter(|value| !value.trim().is_empty());
-
-    #[cfg(unix)]
-    {
-        let uid = unsafe { libc::getuid() };
-        match name {
-            Some(name) => format!("{name}, uid {uid}"),
-            None => format!("uid {uid}"),
-        }
-    }
-
-    #[cfg(not(unix))]
-    {
-        name.unwrap_or_else(|| "unknown user".to_string())
-    }
-}
-
-fn runtime_root_owner_label(runtime_root: &Path) -> String {
-    match fs::metadata(runtime_root) {
-        #[cfg(unix)]
-        Ok(metadata) => format!("Runtime temp root owner: uid {}.", metadata.uid()),
-        #[cfg(not(unix))]
-        Ok(_) => "Runtime temp root exists but ownership could not be inspected on this platform."
-            .to_string(),
-        Err(e) => format!("Runtime temp root owner could not be inspected: {e}."),
     }
 }
